add easyeyesBaseUrl as a module
