+++ components/save-snapshots/boxIntegration.js
@@ -1,4 +1,5 @@
 import { captureError } from "../../../source/sentry";
+import { getEasyEyesBaseUrl } from "../easyeyesBaseUrl";
 
 export const saveSnapshot = async (image, experimentID, participantID) => {
   try {
@@ -28,13 +29,4 @@ export const saveSnapshot = async (image, experimentID, participantID) => {
   }
 };
 
-const getBaseUrl = () => {
-  const urlParams = new URLSearchParams(window.location.search);
-  const previewDeployBase = urlParams.get("preview-deploy");
-
-  if (previewDeployBase) return previewDeployBase;
-  if (window.location.hostname === "localhost") return "http://localhost:8888";
-  return "https://easyeyes.app";
-};
-
-const getBoxApiUrl = () => getBaseUrl() + "/.netlify/functions/box-api";
+const getBoxApiUrl = () => getEasyEyesBaseUrl() + "/.netlify/functions/box-api";
